fixed bug
products weren't being deleted from the list; this is now fixed. One thing still bugs me; modelstate is still invalid when OnPostDelete is called...Other than that biggest issues here was the redirect to the same page; I had to add another parameter to OnGet in order for it to work, that param will not be used, despite containing a value.